WIP: Draw rounded rectangles with anti-aliasing
Draw rounded rectangles using a distance function. This removes the need
to use lyon to tesselate the rounded corners and as a bonus it gives
anti-aliased borders.

There are still rendering issues, the transition between border and fill
is not correct. Also there's code commented out that shouldn't be, and
FIXMEs :)